requirements.txt
pytube
moviepy
pyyaml